Cleanup document isolation, improve tests isolation
